pyproject.toml
[tool.poetry]
name = "smart-inventory-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.14"
instructor = "^1.4.0"
cohere = "^5.8.1"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
